alterando o tipo do atributo salarioBruto e salarioLiquido de BigDecimal para Double.

+++ src/main/java/io/github/patricioor/rh_company/application/dto/FolhaPagamento/FolhaPagamentoDTO.java
@@ -3,7 +3,6 @@ package io.github.patricioor.rh_company.application.dto.FolhaPagamento;
 import lombok.Getter;
 import lombok.Setter;
 
-import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 import java.util.UUID;
@@ -14,8 +13,8 @@ public class FolhaPagamentoDTO {
     private UUID id;
     private LocalDate dataEmissao;
     private String periodoReferencia;
-    private BigDecimal salarioBruto;
-    private BigDecimal salarioLiquido;
+    private Double salarioBruto;
+    private Double salarioLiquido;
     private String funcionarioId;
 
     List<DescontoDTO> descontos;
